COPSPR-7097 - Modifying disk size type to double

+++ azure-to-oci.py
@@ -123,7 +123,7 @@ if __name__ == "__main__":
 
     # create the VM from the imported image
     oci_shape = instance_size
-    oci_disk_size = double(get_vm_config(vm_name)["disk_size"])
+    oci_disk_size = float(get_vm_config(vm_name)["disk_size"])
     oci_disk =  oci_disk_size + (oci_disk_size*(50/100))
     oci_create_vm_from_image(qcow2_file, oci_shape, oci_disk)
